pyproject.toml
[project]
name = "Quiz-Back-FastAPI"
version = "0.0.1"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.124.2",
  "uvicorn>=0.38.0",
  "sqlmodel>=0.0.22",         # SQLAlchemy 2.x + Pydantic v2, parfait pour SQLite
  "pydantic-settings>=2.5",
  "passlib[argon2]>=1.7.4",
  "python-jose[cryptography]>=3.5.0",
]

[tool.ruff]
line-length = 100